replace floats with double

+++ src/01_ProjectileRenderer/Projectile.h
@@ -5,8 +5,8 @@
 class Projectile {
 public:
     Projectile(
-        const float p_x, const float p_y, const float p_z,
-        const float v_x, const float v_y, const float v_z)
+        const double p_x, const double p_y, const double p_z,
+        const double v_x, const double v_y, const double v_z)
     : position{p_x, p_y, p_z}
     , velocity{v_x, v_y, v_z}
     {}
