+++ src/Constvals.h
@@ -1,5 +1,5 @@
 #pragma once
 
-constexpr float EPSILON {1E-4f};
+constexpr double EPSILON {1E-4f};
 
-auto EQF(const float& a, const float& b) -> bool;
+auto EQF(const double& a, const double& b) -> bool;
