+++ src/Ray.cpp
@@ -5,7 +5,7 @@ Ray::Ray(const Point& origin, const Vector& direction)
 , direction{direction}
 {}
 
-auto Ray::Position(const float& t) const -> Point {
+auto Ray::Position(const double& t) const -> Point {
     return origin + direction*t;
 }
 
@@ -13,7 +13,7 @@ auto Ray::Intersects(const Geometry& s) const -> std::vector<Intersection> {
     return s.Intersects(*this);
 }
 
-auto Ray::Transform(const Matrix4f& matrix) const -> Ray {
+auto Ray::Transform(const Matrix4d& matrix) const -> Ray {
     return Ray{
         matrix*origin,
         matrix*direction
