+++ src/Sphere.cpp
@@ -6,7 +6,7 @@
 #include "Point.h"
 #include "Vector.h"
 
-Sphere::Sphere(const Matrix4f& transform, const Material_t& material)
+Sphere::Sphere(const Matrix4d& transform, const Material_t& material)
 : transform{transform}
 , material{material}
 {}
@@ -18,11 +18,11 @@ auto Sphere::Clone() const -> GeometryPtr {
 auto Sphere::operator==(const Sphere& rhs) const -> bool {
     return true;
 }
-auto Sphere::Transform() const -> Matrix4f {
+auto Sphere::Transform() const -> Matrix4d {
     return transform;
 }
 
-auto Sphere::Transform(const Matrix4f& other) -> void {
+auto Sphere::Transform(const Matrix4d& other) -> void {
     transform = other;
 }
 
@@ -63,6 +63,6 @@ auto Sphere::Material(const Material_t& mat) -> void {
     material = mat;
 }
 
-auto SphereFactory(const Matrix4f& transform, const Material_t& material) -> GeometryPtr {
+auto SphereFactory(const Matrix4d& transform, const Material_t& material) -> GeometryPtr {
     return std::make_unique<Sphere>(transform, material);
 }
