+++ src/Transformations.h
@@ -4,17 +4,17 @@
 
 class Transformations {
 public:
-    static auto Translation(const float& tx, const float& ty, const float& tz) -> Matrix4f;
-    static auto Scaling(const float& sx, const float& sy, const float& sz) -> Matrix4f;
-    static auto RotationX(const float& rad) -> Matrix4f;
-    static auto RotationY(const float& rad) -> Matrix4f;
-    static auto RotationZ(const float& rad) -> Matrix4f;
-    static auto Shearing(const float& xy, const float& xz,
-                         const float& yx, const float& yz,
-                         const float& zx, const float& zy) -> Matrix4f;
+    static auto Translation(const double& tx, const double& ty, const double& tz) -> Matrix4d;
+    static auto Scaling(const double& sx, const double& sy, const double& sz) -> Matrix4d;
+    static auto RotationX(const double& rad) -> Matrix4d;
+    static auto RotationY(const double& rad) -> Matrix4d;
+    static auto RotationZ(const double& rad) -> Matrix4d;
+    static auto Shearing(const double& xy, const double& xz,
+                         const double& yx, const double& yz,
+                         const double& zx, const double& zy) -> Matrix4d;
 
 };
 
 class Point;
 class Vector;
-auto ViewTransform(const Point& from, const Point& to, const Vector& up) -> Matrix4f;
+auto ViewTransform(const Point& from, const Point& to, const Vector& up) -> Matrix4d;
